+++ inst/staticdocs/index.R
@@ -162,6 +162,7 @@ sd_section("Counts/Descriptives",
     "termco_c",
     "wfm",
     "word_count",
+    "word_lengths",
     "word_stats"
   )
 )